daemon auth methods
